Refactor: Move animation functions to Vehicle.Animation.cpp
This moves the flat ride operating logic like the swinging, ferris wheel, rotating, simulator, haunted house, top spin, circus, etc etc, swing/spin car physics, and the vehicle animation system for the steam locomotive, swan boat, multi-dimension, animal flying,  etc etc etc to  Vehicle.Animation.cpp. Animates vehicleBreakdown from static to extern to allow shared access across translation units
